third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-unknown-hash-allowed.html

<!DOCTYPE html>
<html>
    <body>
        <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <script src='pass.js' integrity="ni:///sha1;Ywd5te2RvxmuHLzLzZjYo1_0LMw="></script>
        <script src='pass.js' integrity="ni:///#yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8="></script>
	</body>
</html>
