third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-unknown-hash-allowed-expected.txt

CONSOLE ERROR: Error parsing 'integrity' attribute ('ni:///sha1;Ywd5te2RvxmuHLzLzZjYo1_0LMw='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
ALERT: PASS
CONSOLE ERROR: Error parsing 'integrity' attribute ('ni:///#yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
ALERT: PASS
